Added test for storage backup

+++ android-app/app/src/main/java/com/arsvechkarev/vault/features/common/data/StorageBackupFileSaver.kt
@@ -8,15 +8,21 @@ import app.keemobile.kotpass.database.KeePassDatabase
 import app.keemobile.kotpass.database.encode
 import com.arsvechkarev.vault.core.DispatchersFacade
 import com.arsvechkarev.vault.features.common.AppConstants.MIME_TYPE_ALL
+import com.arsvechkarev.vault.features.common.domain.BackupInterceptor
 import com.arsvechkarev.vault.features.common.model.BackupFileData
 import kotlinx.coroutines.withContext
 
 class StorageBackupFileSaver(
   private val context: Context,
-  private val dispatchers: DispatchersFacade
+  private val dispatchers: DispatchersFacade,
+  private val backupInterceptor: BackupInterceptor,
 ) {
   
   suspend fun getAll(directory: String): List<BackupFileData> = withContext(dispatchers.IO) {
+    val interceptedList = backupInterceptor.interceptGetAll()
+    if (interceptedList != null) {
+      return@withContext interceptedList
+    }
     val tree = checkNotNull(DocumentFile.fromTreeUri(context, Uri.parse(directory)))
     tree.listFiles().filter { it.name != null }
         .map { BackupFileData(checkNotNull(it.name), it.lastModified()) }
@@ -26,6 +32,9 @@ class StorageBackupFileSaver(
     directory: String,
     files: List<BackupFileData>
   ) = withContext(dispatchers.IO) {
+    if (backupInterceptor.interceptDeleteAll(files)) {
+      return@withContext
+    }
     files.forEach {
       val backupFileUri = getOrCreateBackupFileUri(Uri.parse(directory), it.name)
       DocumentFile.fromSingleUri(context, backupFileUri)?.delete()
@@ -38,6 +47,9 @@ class StorageBackupFileSaver(
     filename: String,
     database: KeePassDatabase
   ) = withContext(dispatchers.IO) {
+    if (backupInterceptor.interceptBackup(filename, database)) {
+      return@withContext
+    }
     val newBackupFile = getOrCreateBackupFileUri(Uri.parse(directory), filename)
     context.contentResolver.openOutputStream(newBackupFile)?.use(database::encode)
   }
